+++ CHANGELOG.md
@@ -62,7 +62,8 @@ Developers may need to change their code
 
 ### Corrected Bugs
 
-* Fixed: Fixed: memory error in rpc_get2 when error handling
+* Fixed: Restconf notification: Add "data:" as prefix to every line., not just first
+* Fixed: memory error in rpc_get2 when error handling
 * Fixed: [RESTCONF crashes after sending request that ends with &](https://github.com/clicon/clixon/issues/618)
 * Fixed: Do not use d_type from readdir()